Check model estimated total catch to the GEMM estimates. This function should only be used to compare estimates for coastwide models.

Check model estimated total catch to the GEMM estimates. This function should only be used to compare estimates for coastwide models.

compare_model_gemm_catch(
  replist,
  common_name,
  gemm_data = NULL,
  dir = NULL,
  add_name = NULL,
  verbose = TRUE
)

Arguments

replist

List of model objects created by r4ss::SS_output().

common_name

Common name of the species in the GEMM database.

gemm_data

A data frame of GEMM data from nwfscSurvey::pull_gemm(). If this argument is left blank, the function will pull the GEMM data and evaluate mortality on a coastwide basis. However, if your model is not coastwide and the GEMM data can be split out to match the model area (e.g., assessment split north and south of 40 10 N. latitude) this filtered data frame can be passed to the function via this argument and the total mortality comparisons will be done relative to that data object. The default value is NULL.

dir

Directory to save figure. If no directory is provided the figure is not saved. If a dir is specified a figure of the model estimated total mortality compared to the GEMM total mortality, a figure of total mortality across all years for the model, a rda of all model years landed and catch estimates by fleet, and a rda of the model estimated catch compared to the GEMM from 2003+. Default is NULL.

add_name

String that can be added to the saved figure names.

verbose

A logical that specifies if you want to print messages and warnings to the console. The default is TRUE.

Author

Chantel Wetzel