Check that all of the settings are in the list

Check that all of the settings are in the list

get_settings(settings = NULL, verbose = FALSE)

Arguments

settings

A list of the current settings where each object in the list must be named. Those names that are not found in the stored list will be added. The default value of NULL leads to a full list being returned.

verbose

A logical value specifying if the message should be output to the screen or not.

Value

A list of settings for running model diagnostics.

Author

Chantel Wetzel & Kelli Faye Johnson

Examples

get_settings(list("Njitter" = 10))
#> $Njitter
#> [1] 10
#> 
#> $base_name
#> [1] "model"
#> 
#> $para_offset
#> [1] FALSE
#> 
#> $run
#> [1] "jitter"  "profile" "retro"  
#> 
#> $profile_details
#>              parameters   low high step_size param_space
#> 1 NatM_uniform_Fem_GP_1  0.40  0.4      0.01  multiplier
#> 2           SR_BH_steep  0.25  1.0      0.05        real
#> 3             SR_LN(R0) -2.00  2.0      0.25    relative
#> 
#> $version
#> [1] "3.30"
#> 
#> $exe
#> [1] "ss3"
#> 
#> $verbose
#> [1] FALSE
#> 
#> $extras
#> [1] "-nohess"
#> 
#> $show_in_console
#> [1] TRUE
#> 
#> $printlikes
#> [1] FALSE
#> 
#> $jitter_fraction
#> [1] 0.05
#> 
#> $jitter_init_values_src
#> NULL
#> 
#> $oldsubdir
#> [1] ""
#> 
#> $newsubdir
#> [1] "retro"
#> 
#> $retro_yrs
#> [1] -1 -2 -3 -4 -5
#> 
#> $overwrite
#> [1] TRUE
#> 
#> $show_in_console
#> [1] FALSE
#> 
#> $btarg
#> NULL
#> 
#> $minbthresh
#> NULL
#> 
#> $remove_files
#> [1] TRUE
#> 
#> $oldctlfile
#> [1] "control.ss_new"
#> 
#> $newctlfile
#> [1] "control_modified.ss"
#> 
#> $linenum
#> NULL
#> 
#> $string
#> NULL
#> 
#> $profilevec
#> NULL
#> 
#> $usepar
#> [1] FALSE
#> 
#> $globalpar
#> [1] FALSE
#> 
#> $parlinenum
#> NULL
#> 
#> $parstring
#> NULL
#> 
#> $saveoutput
#> [1] TRUE
#> 
#> $overwrite
#> [1] TRUE
#> 
#> $whichruns
#> NULL
#> 
#> $prior_check
#> [1] FALSE
#> 
#> $read_like
#> [1] TRUE
#> 
#> $init_values_src
#> [1] 0
#> 
#> $subplots
#> [1] 1 3
#>