Generate results from likelihood profiles

Used by profile_wrapper() to write summary statistics to the disk.

get_summary(mydir, para, vec, profilemodels, profilesummary)

Arguments

mydir

Directory where model files are located. There is no default entry for mydir. An example of user input could be mydir = file.path("C:/my_models", "base_model").

para

A character string specifying the SS3 parameter name that the profile pertains to. The parameter name should match the name in the control.ss_new file from SS3.

vec

A numeric vector specifying the parameter values that the profile covers.

profilemodels

An object returned from r4ss::SSgetoutput().

profilesummary

An object returned from r4ss::SSsummarize().

Author

Chantel Wetzel & Kelli Johnson