Generate likelihood profile plots

Create plots to be included in presentations and documents that summarize a likelihood profile.

plot_profile(mydir, rep, para, profilesummary)

Arguments

mydir

Directory where model files are located. There is no default entry for mydir. An example of user input could be mydir = file.path("C:/my_models", "base_model").

rep

A list of model output as returned by r4ss::SS_output().

para

A character string specifying the SS3 parameter name that the profile pertains to. The parameter name should match the name in the control.ss_new file from SS3. The name will be cleaned up for plotting purposes internally within the function. For example, SR_BH_steep becomes "Steepness (h)".

profilesummary

Output from r4ss::SSsummarize(). Ensure that the results are ordered according to the parameter of interest. Otherwise, the profile plots will not be smooth and lines will zig-zag back and forth.

See also

Author

Chantel Wetzel.