Run r4ss::profile() based on model_settings

Setting up the specifications, running the profile using r4ss::profile(), and generating figures and tables can be tedious, error prone, and time consuming. Thus, profile_wrapper() aims to further decrease the work needed to generate a profile that is easily included in a assessment report for the Pacific Fisheries Management Council. See the See Also section for information on a workflow to use this function.

profile_wrapper(mydir, model_settings)

Arguments

mydir

Directory where model files are located. There is no default entry for mydir. An example of user input could be mydir = file.path("C:/my_models", "base_model").

model_settings

A list of all model settings. Typically, created using get_settings but users can create their own list, as long as it has all of the necessary components for the function at hand.

Value

Nothing is explicitly returned from profile_wrapper(). The following objects are saved to the disk:

*_profile_output.Rdata

*_trajectories_* from r4ss::SSplotComparisons()

piner_panel_*.png

parameter_panel_*.png

run_diag_warning.txt

a copy of the control file saved to model_settings$newctlfile

backup_oldctlfile.ss

backup_ss.par

See also

Author

Chantel Wetzel and Ian Taylor.