Run MCMC using adnuts for SS3

Run MCMC using adnuts for SS3

run_mcmc_diagnostics(
  dir_wd,
  model = "ss3",
  extension = ".exe",
  iter = 2000,
  chains = 2,
  interactive = FALSE,
  verbose = FALSE
)

Arguments

dir_wd

A file path to the directory where the model files and the executable are saved. This can be relative to your current working directory or a full path. Your working directory will be changed while executing the function to be inside the location provided in this argument but upon exiting it will return to your current working directory.

model

A string giving the name of the executable without an an extension. The extension is provided in extension. Do not include any path information, just the name of the executable. For example, the default is "ss3".

extension

A string giving the extension for the executable name provided in model. The default is .exe, which works on Windows machines.

iter

A numeric value for the number of MCMC draws to do. Default is 2000.

chains

A numeric value for the number of chains to run. Default is 2.

interactive

A logical, where TRUE will run adnuts::launch_shinyadmb(). The default is FALSE.

verbose

A logical, specifying if information should be printed to the screen. The default is FALSE but using TRUE can be helpful when you are first running the function or when you have a very unstable model.

Author

Cole M. Monnahan (cole.monnahan@noaa.gov; February 2021) and augmented by Kelli F. Johnson